Add baud rate control (BRC) pin support, currently hardcoded to pin 2. This makes the Roomba wake up from sleep, so you don't need to push the button on wakeup.
Also added "was_disconnected" code, to automatically reconnect if the Arduino
self-disconnected.  This fixes the bug where a transient network delay would hang up the
backend, causing the Arduino to briefly lose pc_connected and turn off the Roomba.
Now the Arduino turns the Roomba back on when this happens (to test: ctrl-Z backend, wait, fg)