python3 is installed on ubuntu 14.04 by default
Futhermore, its a oneliner to install on most 'serious' OSX systems.
Windows 10 requires a download for both 2 and 3, so the point is moot
there.

For ubuntu, please see:
- http://askubuntu.com/a/449558

Reasons for:
    - range() isnt literally the worst implemenation of range() ever
    - Going into the modern age
        - Many libraries which we might want in the future have stopped
          writing for python2.
        - Python2 is only getting bare bones security updates
    - If we want to thread this later, python3 thread support is much
        better
    - Much better unicode support

Revert "Python3 is not installed by default on OSX, Ubuntu 14.04, or on Windows 10... Why not stick with 2 since that is installed on everything by default..."

This reverts commit 6ef075ae9accddecd2520bd7f14b3e11ee1dda4f.